SolidFire snapshot rollback
No guarantee or warranty of usage are offered or provided. Use at your own risk.

This is designed to allow you to submit a set of volumes and have the script manage group snapshots.  


python test2.py -srcmv sfdemo1 -dstmv sfdemo2 -srcun admin -dstun admin -srcpw Netapp1! -dstpw Netapp1! -rt 00:60:00 –va 2 3 4

-srcmv: The source cluster MVIP
-dstmv: The destination cluster MVIP
-srcun: The source cluster username
-dstun: The destination cluster username
-srcpw: The source cluster user’s password
-dstpw: The destination cluster user’s password
-rt: The retention time of the snapshot, format is HH:mm:ss, enter Zero’s if leaving any section blank
-va: The list of volume IDs to add, should be space separated integers

The volume IDs that should be taken as a consistent group, cluster names, and authorization information are submitted.
From there the script logs in to the source cluster
It creates a group snapshot on the volumes submitted
It counts the volumes to know how many snapshots should be in the group
It looks for the member snaps in that group
It builds an array of snapshot UUIDs and stores them
It then verifies the snap is on the remote cluster
It pauses until the snaps are reported on both clusters

Once snapshots are verified on the remote the script connects to the destination cluster
Looks at the snapshots and deletes any named “rollback” (see below for rollback snapshots
Verifies the snapshot name and creates an array of the snapshots
Compares the number in the destination array to the number the source found and exits out if they do not match
Loops through the volumes and verifies the volumes are in an “Idle” state for transfers
Verifies the snapshot UUIDs on the destination match the source
Pauses the transfer by setting the volume to readWrite
Rolls the snapshot into the volume
Creates a snapshot called “rollback” in case the snapshot update doesn’t work on all volumes
Resets the volume to replicationTarget to allow replication to resume

Additional error checking is needed
Volumes that are already paused/readWrite – what do we do?
Need to verify all volumes in group have snapshot applied or revert to previous state using rollback snapshot
Should probably ensure that the clusters are paired
The script has a limitation that the source and destination volume IDs match, I am working on this currently.  Easier to PoC with matching sets, in the real world this will hardly ever be the case.  Fixed 16-Mar-2018
Need to verify and retry if a volume does not revert to replicationTarget from readWrite
Logging needs to be set to actual logging rather than simply printing to the terminal
Consolidate volumes to be moved to readWrite and back to replicationTarget as a group instead of one at a time